Extract workflow create/run into lib/workflow_utils
Moves replace_uuids, get_task_signature, create_workflow_signature, and
the module-level Celery app out of api/v1/workflows.py into
lib/workflow_utils.py, and adds two new public helpers:

- create_workflow_from_template(): resolve template (if any), render
  params, create a result subfolder, and persist the workflow row.
- run_workflow(): persist the spec, build the celery canvas, and dispatch.

The HTTP routes in api/v1/workflows.py become thin shells that delegate
to these helpers. This opens the door callers (e.g. importer) to invoke workflows without using the REST API.

Also introduces lib/workflow_spec_utils.py for pure-Python workflow spec
manipulation (add_unique_parameter_names). Splitting this out of
workflow_utils.py avoids a circular import between crud/workflow and
lib/workflow_utils.
