+++ src/lib/workflow_utils.py
@@ -1,4 +1,4 @@
-# Copyright 2025 Google LLC
+# Copyright 2025-2026 Google LLC
 #
 # Licensed under the Apache License, Version 2.0 (the "License");
 # you may not use this file except in compliance with the License.
@@ -12,7 +12,41 @@
 # See the License for the specific language governing permissions and
 # limitations under the License.
 
-from collections import defaultdict
+"""Workflow helpers reusable by the HTTP layer and in-process callers.
+
+These functions do not depend on FastAPI; they exist so that both
+``api.v1.workflows`` (HTTP) and, in the future, the bundled importers can
+create and run workflows without going through the REST API.
+"""
+
+import json
+import os
+from typing import Optional
+from uuid import uuid4
+
+from celery import chain as celery_chain
+from celery import chord as celery_chord
+from celery import group as celery_group
+from celery import signature
+from celery.app import Celery
+from celery.canvas import Signature
+from sqlalchemy.orm import Session
+
+from api.v1 import schemas
+from datastores.sql.crud.folder import create_subfolder_in_db
+from datastores.sql.crud.workflow import (
+    create_task_in_db,
+    create_workflow_in_db,
+    get_workflow_template_from_db,
+)
+from datastores.sql.models.workflow import Task, Workflow
+
+
+# Celery app shared by everything that dispatches workflow signatures.
+# Constructed at import time to match the behavior of the previous module-level
+# instance in ``api.v1.workflows``.
+_redis_url = os.getenv("REDIS_URL")
+celery_app = Celery(broker=_redis_url, backend=_redis_url)
 
 
 def update_task_config_values(data: dict | list, parameters: dict) -> None:
@@ -43,39 +77,310 @@ def update_task_config_values(data: dict | list, parameters: dict) -> None:
             update_task_config_values(item, parameters)
 
 
-def add_unique_parameter_names(data: dict | list) -> None:
-    """
-    Wrapper function to initiate the recursive traversal with a shared counter.
+def get_task_signature(
+    db: Session,
+    current_user: schemas.User,
+    task_data: dict,
+    input_files: list,
+    output_path: str,
+    workflow: schemas.Workflow,
+) -> Signature:
+    """Returns a Celery task signature for a given task.
 
     Args:
-        data (dict | list): The workflow dictionary to modify.
+        db (Session): The database session.
+        current_user (schemas.User): The current user.
+        task_data (dict): The task data.
+        input_files (list): A list of input files.
+        output_path (str): The output path.
+        workflow (schemas.Workflow): The workflow.
+
+    Returns:
+        Signature: The Celery task signature.
     """
-    _add_unique_parameter_names_recursive(data, defaultdict(int))
+    task_uuid = task_data.get("uuid", uuid4().hex)
+    task_config = {
+        option["name"]: option.get("value") for option in task_data.get("task_config", {})
+    }
+
+    # Create a new DB task
+    new_task_db = Task(
+        display_name=task_data.get("display_name"),
+        description=task_data.get("description"),
+        config=json.dumps(task_config),
+        uuid=task_uuid,
+        user=current_user,
+        workflow=workflow,
+    )
+    create_task_in_db(db, new_task_db)
+
+    task_signature = signature(
+        task_data.get("task_name"),
+        kwargs={
+            "input_files": input_files,
+            "output_path": output_path,
+            "workflow_id": workflow.id,
+            "task_config": task_config,
+        },
+        queue=task_data.get("queue_name"),
+        task_id=task_uuid,
+    )
+    return task_signature
 
 
-def _add_unique_parameter_names_recursive(data: dict | list, counts: defaultdict) -> None:
-    """
-    Recursively traverses a workflow dictionary and adds a unique
-    "param_name" key to each task_config item that has a "name" key.
-    The "param_name" is created by lowercasing the original name and
-    replacing spaces with underscores, followed by a unique index.
+def create_workflow_signature(
+    db: Session,
+    current_user: schemas.User,
+    task_data: dict,
+    input_files: list,
+    output_path: str,
+    workflow: schemas.Workflow,
+) -> Signature:
+    """Creates a Celery workflow signature for a given task definition
+
+    This function recursively constructs a Celery workflow signature based on the
+    provided `task_data`, which represents a structured description of tasks and their
+    dependencies. It supports two primary task types: 'chain' and 'task'.
+
+    chain: Represents a sequence of tasks executed in order.
+        -   If the chain contains multiple tasks, a Celery `celery_group` is created to
+            execute them concurrently. celery_group allows multiple tasks to be run
+            in parallel.
+        -   If only one task is present, a Celery `celery_chain` is created to execute
+            it *serially*. `celery_chain` ensures tasks are executed one after another,
+            ith the output of one task becoming the input of the next.
+
+    task: Represents a single, executable task.
+        - It retrieves the corresponding Celery task signature using get_task_signature.
+        - If the task has sub-tasks, they are incorporated into the workflow using
+          Celery `celery_chain` and `celery_group` constructs, depending on the number
+          of sub-tasks. The primary task is chained with the subtasks.
+
+    The function effectively translates a hierarchical task description into a Celery
+    workflow that can be executed asynchronously. This allows for complex workflows to
+    be defined and executed in a distributed manner.
 
     Args:
-        data (dict | list): The workflow dictionary to modify.
-        counts (defaultdict): A dictionary to keep track of the counts of each normalized name.
+        db (Session): The database session.
+        current_user (schemas.User): The current user.
+        task_data (dict): The task data.
+        input_files (list): A list of input files.
+        output_path (str): The output path.
+        workflow (schemas.Workflow): The workflow.
+
+    Returns:
+        Signature: The Celery workflow signature.
+
+    Raises:
+        ValueError: If the task type is not supported.
+    """
+    if task_data["type"] == "chain":
+        if len(task_data["tasks"]) > 1:
+            return celery_group(
+                create_workflow_signature(
+                    db, current_user, task, input_files, output_path, workflow
+                )
+                for task in task_data["tasks"]
+            )
+        else:
+            return celery_chain(
+                create_workflow_signature(
+                    db,
+                    current_user,
+                    task_data["tasks"][0],
+                    input_files,
+                    output_path,
+                    workflow,
+                )
+            )
+
+    elif task_data["type"] == "chord":
+        header_tasks = [
+            create_workflow_signature(db, current_user, t, input_files, output_path, workflow)
+            for t in task_data.get("tasks", [])
+        ]
+
+        callback_task_data = task_data.get("callback")
+        if not callback_task_data:
+            raise ValueError("Chord definition requires a 'callback' task.")
+
+        callback_signature = create_workflow_signature(
+            db, current_user, callback_task_data, input_files, output_path, workflow
+        )
+
+        return celery_chord(header_tasks, callback_signature)
+
+    elif task_data["type"] == "task":
+        task_signature = get_task_signature(
+            db, current_user, task_data, input_files, output_path, workflow
+        )
+        if task_data["tasks"]:
+            if len(task_data["tasks"]) > 1:
+                return celery_chain(
+                    task_signature,
+                    celery_group(
+                        create_workflow_signature(
+                            db, current_user, t, input_files, output_path, workflow
+                        )
+                        for t in task_data["tasks"]
+                    ),
+                )
+            else:
+                return celery_chain(
+                    task_signature,
+                    create_workflow_signature(
+                        db,
+                        current_user,
+                        task_data["tasks"][0],
+                        input_files,
+                        output_path,
+                        workflow,
+                    ),
+                )
+        else:
+            return task_signature
+    else:
+        raise ValueError(f"Unsupported task type: {task_data['type']}")
+
+
+def replace_uuids(data: dict | list, replace_with: str = None) -> dict | list:
+    """Recursively replaces UUID keys within a dictionary or list structure.
+
+    This function traverses the provided `data` structure (which can be a dictionary or
+    list) and replaces any dictionary keys named "uuid" with a new value.
+
+    If `replace_with` is not provided (or is None), a newly generated UUID is used as
+    the replacement. If `replace_with` is provided, that value is used as the
+    replacement for all "uuid" keys.
+
+    This is needed when modifying workflow specifications that contain UUIDs, ensuring
+    that each instance has unique identifiers.
+
+    Args:
+        data (dict | list): The dictionary or list to traverse and modify.
+        replace_with (str, optional): The value to replace UUIDs with. Defaults to None.
+
+    Returns:
+        dict | list: The modified dictionary or list with UUIDs replaced.
     """
     if isinstance(data, dict):
-        if "task_config" in data and isinstance(data["task_config"], list):
-            for item in data["task_config"]:
-                if "name" in item:
-                    normalized_name = item["name"].lower().replace(" ", "_")
-                    param_name = f"{normalized_name}_{counts[normalized_name]}"
-                    item["param_name"] = param_name
-                    counts[normalized_name] += 1
-
         for key, value in data.items():
-            _add_unique_parameter_names_recursive(value, counts)
-
+            if key == "uuid":
+                if not replace_with:
+                    data[key] = uuid4().hex
+                else:
+                    data[key] = replace_with
+            else:
+                replace_uuids(value)
     elif isinstance(data, list):
         for item in data:
-            _add_unique_parameter_names_recursive(item, counts)
+            replace_uuids(item)
+
+
+class TemplateNotFoundError(ValueError):
+    """Raised when ``create_workflow_from_template`` cannot resolve a template id."""
+
+
+def create_workflow_from_template(
+    db: Session,
+    *,
+    folder_id: int,
+    file_ids: list[int],
+    template_id: Optional[int],
+    template_params: Optional[dict],
+    user: schemas.User,
+) -> Workflow:
+    """Create a workflow row (optionally from a template) and return it.
+
+    Mirrors the behavior of the HTTP ``POST /folders/{folder_id}/workflows/``
+    route minus the request-parsing shell. No authorization is performed;
+    the caller is trusted (either an authenticated HTTP route that has
+    already checked access, or an in-process invoker such as an importer).
+
+    Raises:
+        TemplateNotFoundError: When ``template_id`` is provided but the
+            template does not exist.
+    """
+    default_workflow_display_name = "Untitled workflow"
+    default_spec_json = None
+    from_template = None
+
+    if template_id:
+        from_template = get_workflow_template_from_db(db, template_id)
+        if not from_template:
+            raise TemplateNotFoundError(
+                f"Workflow template {template_id} not found"
+            )
+        default_workflow_display_name = from_template.display_name
+        spec_json = json.loads(from_template.spec_json)
+        # Replace the placeholder UUIDs seeded in the template with fresh ones
+        # so each workflow instance has unique identifiers.
+        replace_uuids(spec_json)
+        if template_params:
+            update_task_config_values(spec_json, template_params)
+        default_spec_json = json.dumps(spec_json)
+
+    # Create a new folder to hold workflow results.
+    new_folder = schemas.FolderCreateRequest(
+        display_name=default_workflow_display_name, parent_id=folder_id
+    )
+    new_workflow_folder = create_subfolder_in_db(db, folder_id, new_folder, user)
+
+    new_workflow_db = schemas.Workflow(
+        display_name=default_workflow_display_name,
+        user_id=user.id,
+        spec_json=default_spec_json,
+        file_ids=file_ids,
+        folder_id=new_workflow_folder.id,
+        template_id=from_template.id if from_template else None,
+    )
+    return create_workflow_in_db(db, new_workflow_db)
+
+
+def run_workflow(
+    db: Session,
+    *,
+    workflow: Workflow,
+    workflow_spec: dict,
+    user: schemas.User,
+) -> Workflow:
+    """Persist ``workflow_spec`` on ``workflow`` and dispatch it via Celery.
+
+    Builds input-file dicts from ``workflow.files``, ensures the output
+    directory exists, constructs the Celery canvas, and calls
+    ``apply_async()``. Returns the refreshed workflow row.
+    """
+    workflow.spec_json = json.dumps(workflow_spec)
+
+    input_files = [
+        {
+            "id": file.id,
+            "uuid": file.uuid.hex,
+            "display_name": file.display_name,
+            "extension": file.extension,
+            "data_type": file.data_type,
+            "mime_type": file.magic_mime,
+            "path": file.path,
+        }
+        for file in workflow.files
+    ]
+
+    output_path = workflow.folder.path
+    if not os.path.exists(output_path):
+        os.makedirs(output_path)
+
+    celery_workflow = create_workflow_signature(
+        db,
+        user,
+        workflow_spec.get("workflow"),
+        input_files,
+        output_path,
+        workflow,
+    )
+    celery_workflow.apply_async()
+
+    db.add(workflow)
+    db.commit()
+    db.refresh(workflow)
+
+    return workflow
